# api.yaml
openapi: 3.0.3
info:
  title: WireGuard Obfuscator Easy API
  description: |
    REST API for managing WireGuard VPN server with obfuscation support.
    Provides endpoints for authentication, client management, server configuration,
    statistics, and monitoring.
  version: 1.0.0
  contact:
    name: Alexey Cluster
    email: cluster@cluster.wtf
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: /{webPrefix}
    description: Production server with web prefix
    variables:
      webPrefix:
        default: ""
        description: Optional web prefix (e.g., /vpn)
...
paths:
...
  /api/clients:
    get:
      tags:
        - Clients
      summary: List all clients
      description: Get list of all WireGuard clients with connection statistics
      operationId: getClients
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Clients
      summary: Create new client
      description: Create a new WireGuard client
      operationId: createClient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  description: Client username
                obfuscation:
                  type: boolean
                  default: true
                  description: Enable obfuscation for this client
                enabled:
                  type: boolean
                  default: true
                  description: Enable client
      responses:
        '201':
          description: Client created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/clients/{username}:
    get:
      tags:
        - Clients
      summary: Get client details
      description: Get detailed information about a specific client
      operationId: getClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Client information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Clients
      summary: Update client
      description: Update client configuration
      operationId: updateClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allowed_ips:
                  type: array
                  items:
                    type: string
                    pattern: '^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}/[0-9]{1,2}$'
                  description: Allowed IP addresses in CIDR format
                obfuscator_port:
                  type: integer
                  nullable: true
                  minimum: 1
                  maximum: 65535
                  description: Custom obfuscator port (null to use default)
                masking_type_override:
                  type: string
                  nullable: true
                  enum: [none, obfs4, scramblesuit]
                  description: Override masking type for this client (null to use server default)
                verbosity_level:
                  type: string
                  enum: [0, 1, 2, 3, 4]
                  description: Verbosity level (0-4)
                enabled:
                  type: boolean
                  description: Enable/disable client
      responses:
        '200':
          description: Client updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Clients
      summary: Delete client
      description: Delete a WireGuard client
      operationId: deleteClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Client deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/clients/{username}/regenerate-keys:
    post:
      tags:
        - Clients
      summary: Regenerate client keys
      description: Generate new key pair for a client
      operationId: regenerateClientKeys
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Keys regenerated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  private_key:
                    type: string
                  public_key:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/clients/{username}/config/wireguard:
    get:
      tags:
        - Clients
      summary: Get WireGuard configuration
      description: Download WireGuard configuration file for a client
      operationId: getClientWireGuardConfig
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: WireGuard configuration file
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/clients/{username}/config/obfuscator:
    get:
      tags:
        - Clients
      summary: Get obfuscator configuration
      description: Download obfuscator configuration file for a client
      operationId: getClientObfuscatorConfig
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Obfuscator configuration file
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Obfuscation is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication. For most endpoints, use JWT token obtained from the /api/auth/login endpoint.
        For metrics endpoints (/api/metrics/*), you can use either a JWT token or the dedicated metrics token (obtained from /api/metrics/token).
        Include the token in the Authorization header as: "Bearer <token>"
  parameters:
    Username:
      name: username
      in: path
      required: true
      description: Client username
      schema:
        type: string
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    Client:
      type: object
      properties:
        username:
          type: string
        public_key:
          type: string
        private_key:
          type: string
        allowed_ips:
          type: array
          items:
            type: string
          description: Allowed IP addresses in CIDR format
        obfuscator_port:
          type: integer
          nullable: true
          description: Custom obfuscator port
        masking_type_override:
          type: string
          nullable: true
          enum: [none, obfs4, scramblesuit]
          description: Override masking type for this client
        verbosity_level:
          type: string
          enum: [0, 1, 2, 3, 4]
          description: Verbosity level
        enabled:
          type: boolean
          description: Whether client is enabled
        is_connected:
          type: boolean
          description: Whether client is currently connected
        latest_handshake:
          type: integer
          description: Unix timestamp of last handshake
        all_time_rx_bytes:
          type: integer
          description: Total received bytes (all time)
        all_time_tx_bytes:
          type: integer
          description: Total sent bytes (all time)
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'